-Updated Section with controller, dao, service and mapper -Working in POSTMAN

+++ src/main/java/com/controlcenter/controlcenter/model/Section.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 public class Section {
     private Long section_id;
     private String section_name;
-    private Long dept_id;
+    private int dept_id;
     private int del_flag;
     private String reg_id;
     private Timestamp reg_date;
